Moved to separated class
